# OpenMP based CPU path optimizations for pooling layer